Prevent blocking connect attempt before server start
  `Server#boot' calls `#responsive?' before starting the server, which
will make the program block on a connection attempt to the host and port
where the server should be running.

  After some time, the `connect' system call returns with an error,
`#responsive?' returns false and `#boot' execution resumes. The server
is then started, `#responsive?' is called again and a new connection
attempt is made.

  This change add an expression in the guard clause for `#responsive?'
method to explicitly test that @server_thread is falsy, and return false
in that case.